vc_rpmsg_service: Userspace interface to stateful VideoCore services
The vc_rpmsg_service driver communicates with VC over the "hostexport"
channel to determine which services to export, and then creates a char
device for each one. Clients can open the char device and then send
and receive messages with the protocol defined in vc_rpmsg_service.h,
with support for passing shared-memory buffers to VC and automatically
cleaning up all resources when the char device is closed.

Signed-off-by: Luke Diamand <luked@broadcom.com>